audo/requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pywinctl>=0.0.42
